+++ test/config.tests.js
@@ -1039,36 +1039,6 @@ describe('Configuration', () => {
         );
       });
 
-      it('should prefix bindingKey with replyTo uuid', () => {
-        configure(
-          {
-            vhosts: {
-              v1: {
-                queues: {
-                  q1: {
-                    replyTo: true,
-                  },
-                },
-                bindings: {
-                  b1: {
-                    source: 'e1',
-                    destination: 'q1',
-                    destinationType: 'queue',
-                    bindingKey: 'foo.bar.#',
-                  },
-                },
-              },
-            },
-          },
-          (err, config) => {
-            assert.ifError(err);
-            assert.strictEqual(config.vhosts.v1.bindings.b1.source, 'e1');
-            assert.strictEqual(config.vhosts.v1.bindings.b1.destination, 'q1');
-            assert.ok(/\w+-\w+-\w+-\w+-\w+\.foo\.bar\.#/.test(config.vhosts.v1.bindings.b1.bindingKey), format('%s failed to match expected pattern', config.vhosts.v1.bindings.b1.bindingKey));
-          }
-        );
-      });
-
       it('should configure multiple bindings from an array of binding keys', () => {
         configure(
           {
@@ -1627,6 +1597,56 @@ describe('Configuration', () => {
           assert.strictEqual(config.publications.p1.encryption.ivLength, 16);
           assert.strictEqual(config.publications.p1.encryption.algorithm, 'algo');
           done();
+        }
+      );
+    });
+
+    it('should suffix referenced replyTo queue with uuid', () => {
+      configure(
+        {
+          vhosts: {
+            v1: {
+              exchanges: ['e1'],
+              queues: {
+                rq: {
+                  replyTo: true,
+                },
+              },
+              publications: {
+                p1: {
+                  exchange: 'e1',
+                  replyTo: 'rq',
+                },
+              },
+            },
+          },
+        },
+        (err, config) => {
+          assert.ifError(err);
+          assert.ok(!config.vhosts.v1.publications);
+          assert.ok(/rq:\w+-\w+-\w+-\w+-\w+/, config.publications.p1.replyTo);
+        }
+      );
+    });
+
+    it('should report unknown replyTo queues', () => {
+      configure(
+        {
+          vhosts: {
+            v1: {
+              exchanges: ['e1'],
+              queues: {},
+              publications: {
+                p1: {
+                  exchange: 'e1',
+                  replyTo: 'rq',
+                },
+              },
+            },
+          },
+        },
+        (err) => {
+          assert.strictEqual(err.message, 'Publication: p1 refers to an unknown reply queue: rq');
         }
       );
     });